test to_dictt for Series
